moved variables to there own folder
