+++ src/main/java/com/aqualen/vsu/trueSkill/FactorGraphs/variable/DefaultVariable.java
@@ -1,4 +1,4 @@
-package com.aqualen.vsu.trueSkill.FactorGraphs;
+package com.aqualen.vsu.trueSkill.FactorGraphs.variable;
 
 public class DefaultVariable<TValue> extends Variable<TValue> {
     public DefaultVariable(TValue value) {
